EDMT-244 refactor : hooks api 폴더 구조 분리에 따른 컴포넌트 코드 경로 수정
