+++ src/components/modal/create-record-modal.tsx
@@ -3,7 +3,7 @@
 import { useRef, useState } from 'react';
 
 import { Input } from '@/components/input/input';
-import { useCreateRecords } from '@/hooks/api/use-create-records';
+import { useCreateRecords } from '@/hooks/api/student-manage/use-create-records';
 import type { RecordType } from '@/types/api/student-record';
 
 import {
