+++ src/components/notice/write-notice.tsx
@@ -6,7 +6,7 @@ import { useRouter } from 'next/navigation';
 
 import TipTapEditor, { type TipTapEditorRef } from '@/components/editor/tiptap-editor';
 import { Input } from '@/components/input/input';
-import { usePostAdminNotice } from '@/hooks/api/use-post-admin-notice';
+import { usePostAdminNotice } from '@/hooks/api/notice/use-post-admin-notice';
 import { revalidateNotice } from '@/lib/actions/revalidateNotice';
 
 const baseStyle = 'rounded-full px-5 py-1 pt-1.5 text-center font-bold text-[14px]';
