+++ src/components/record/record-body.tsx
@@ -5,7 +5,7 @@ import { useState } from 'react';
 import NotAuthorizedError from '@/components/error/not-authorized-error';
 import NotPermissionError from '@/components/error/not-permission-error';
 import Loading from '@/components/loading/loading';
-import { useGetRecords } from '@/hooks/api/use-get-records';
+import { useGetRecords } from '@/hooks/api/student-manage/use-get-records';
 import type { RecordType } from '@/types/api/student-record';
 
 import RecordDetailAdd from './record-detail-add';
