+++ src/components/student-record-write/record-summary.tsx
@@ -5,8 +5,8 @@ import { useRouter } from 'next/navigation';
 import Loading from '@/components/loading/loading';
 import SaveSummaryRecordModal from '@/components/modal/save-summary-record-modal';
 import { Textarea } from '@/components/textarea/textarea';
-import { useGetSummaryRecordDetail } from '@/hooks/api/use-get-summary-record-detail';
-import { usePostSummaryRecordDetail } from '@/hooks/api/use-post-summary-record-detail';
+import { useGetSummaryRecordDetail } from '@/hooks/api/write-record/use-get-summary-record-detail';
+import { usePostSummaryRecordDetail } from '@/hooks/api/write-record/use-post-summary-record-detail';
 import { useAutoResizeTextarea } from '@/hooks/use-auto-resize-textarea';
 import type { RecordType } from '@/types/api/student-record';
 import { calculateByte } from '@/util/calculate-byte';
